Fix bug #386: broken source distribution.
The recent change to do more robust evaluation of int operations (commit
6e3e60f) broke installation of the library grades from the source distribution
on 64-bit machines.  The problem is that we set the target bits-per-word to 32
when building the source distribution.  This means that any evaluation of int
operations done while building the source distribution's C files treats ints as
being 32 bit values.  However, for the source distribution the size of an int
will not be decided until the configure script is run, which is *after* the C
files have been generated.

This change disables all compile time evaluation of int operations when
compiling in .pregen grades.

compiler/const_prop.m:
compiler/simplify_goal_call.m:
	As above.