Fix Java launcher scripts for MSYS.
The shell scripts that are generated for launching programs compiled in the
Java grade are not correct on systems using the MSYS shell.  In particular,
although that shell is mostly a POSIX style shell, we need to interact with the
Windows versions of the Java toolchain and the latter only understands things
like Windows style paths.  This diff generates Java launcher shell scripts for
MSYS that containing paths in the appropriate format, along with any character
escapes required by the shell.

compiler/module_cmds.m:
	As above.