Fix another bug with the Java grade, --use-grade-subdirs and Windows.
When --use-grade-subdirs was enabled the compiler was incorrectly looking for
wrapper batch files generated in the Java grade in the 'bin' directory rather
than the 'bats' directory.  This was due to the file extension for the wrapper
scripts not being set correctly; the fix is to set it correctly.
(Patch from Sebastian Godelet.)

compiler/compile_target_code.m:
	As above.