Simplify the code of recompilation.versions.m.
Remove the predicate item_to_maybe_item_id, which handled all item kinds,
but was used on only a few kinds. Move the code for those kinds to where
they are needed.

Merge some predicates together, because the code interfacing them
was bigger than the smaller predicates themselves.

Mark a likely bug: we don't compare promise items in interfaces.

Mark a possible bug: we don't compare mode_decl items in interfaces if
(due to the use of with_inst) we don't know if they are for a predicate
or a function.

Throw exceptions when we find an item in an interface that doesn't belong
in an interface.