Improve tests/run_one_test.
Replace $x references to shell variables with ${x} references to protect
against mistakes in the extent of the variable name, and put those references
into double quotes when possible to protect against mistakes involving
spaces in variable values or empty strings as variable values.

Initialize local shell variables just before their first use.

Remove the "set -x", since the shell commands that it causes to echo
are just about never interesting, because they are boilerplate code,
the same code being executed for just about every test case, the only
variation being between test cases with and without a -nodepend suffix.
(The interesting shell commands, those which are sometimes needed to
track down the causes of test case failures, are the commands issued by mmake,
which are unaffected by set -x.)

For me, this reduced the size of the output of a bootcheck from 73k+ lines
to 48k+ lines.