Disable parallel marking on Windows.
Disable Boehm GC parallel marking support on Windows for now as it is
broken on Windows 7+.  The problem is fixed in bdwgc commit 57cc049,
whose log message states:

    NT SignalObjectAndWait is broken in Windows 7+ leading to a
    deadlock sometimes because the function is no longer atomic

Alternatively, we could define DONT_USE_SIGNALANDWAIT to force an
implementation that does not use SignalObjectAndWait; that is
essentially what commit 57cc049 does.  However, there might be more
fixes required and we should be upgrading the GC anyway.

configure.ac:
	As above.