Report an error if the compiler is invoked with conflicting modes of operation.
The compiler has a set of options that are supposed to be mutually exclusive,
each of which corresponds to a particular mode of operation, for example
'--make' and '--generate-source-file-mapping'.  Currently, if multiple such
modes of operation are specified on the command line, the one that is actually
performed depends upon the order of the big if-then-else at the beginning of
the predicate main_after_setup/8.  Additionally, there is no feedback to the
user that any other options have been ignored.

This diff changes the compiler to gather up all the modes of operation implied
by the command line options into a set and report an error if there is more
than one element in that set.  In the case where the set contains no elements,
the compiler carries out its default action, as now.  The '--help' and
'--version' options are not included in this set and still take precedence over
any other command line options.

compiler/mercury_compile.m:
	Make the above change; as part of it separate out most of the predicate
	main_after_setup/8 into separate predicates.

	Group an auxiliary predicate together with the code that actually
	uses it.

	Unrelated change: fix spelling in a progress message.