Delete unused test case hard_coded/minint_bug.m
This test case was disabled during a window of time when the behaviour of
string.base_string_to_int changed, such that it fails if the base is 10 and
the integer is int.min_int.

The original behaviour was restored in
commit 8d365b405660c1d40bcb66de955fe6d1d38df4af
and tested as part of general/base_string_to_int_test.m.
minint_bug.m was not re-enabled.

tests/hard_coded/Mmakefile:
tests/hard_coded/minint_bug.m:
	Delete unused and redundant test case.