Improve multi_map.m.
Change the wording of the comments explaining predicates to be more
descriptive, as well as more consistent.

Consistently use K and V as the type variables for the keys and values.

Change variable names to be more consistent.

Use named aux predicates instead of lambdas.

Use cords to avoid reversing and unreversing lists when converting a multi_map
to a flat assoc list.

Avoid lots of memory allocations when counting all the values in a multi_map.

Put related predicates next to each other.