Workaround a problem with 'mmc --make' in the Java grade.
Workaround a problem with 'mmc --make' when the Mercury compiler is built in
the Java grade.  In that grade io.progname will return the program name as
'top_level' because that is what the name of the top-level compiler module is
(in the Java grade the program name is set in the top-level module).  This
means that when '--make' attempts to invoke the compiler that it tries to
invoke the executable 'top_level', which won't exist because we rename it to
'mercury_compile'.

compiler/make.module_target.m:
    If the compiler is built in the Java grade then don't attempt to invoke
    it using the name returned by io.progname, instead just use the value
    of the environment variable MERCURY_COMPILER or, if that is not set, just
    'mmc'.

    Simplify how we generate an error message.