Fix spawn_context bug
In low-level C grades spawn_context attempts to use the stack of the new
context to pass the closure and thread ID to the new context once it starts
running.  The first stack slot used has a 0 offset from the stack pointer,
but the second used has a -1 offset.  If the stack pointer is at the
beginning of the memory zone (which happens sometimes but not usually, see
MR_next_offset()) then this can create an invalid memory access.  This
raises a segfault with the new version of Boehm GC.

Create a stack frame for the closure and thread ID in spawn_context.

library/thread.m:
    As above.