Don't generate near-duplicate tags.
Lots of Mercury types have a name that is also the name of one of the type's
function symbols, like this:

:- type allow_lco
    --->    do_not_allow_lco
    ;       allow_lco.

When jumping to a tag that has two definitions, the type name and one of that
type's function symbols, it doesn't really matter which one we jump to, since
they are next to each other. It is therefore annoying having vim ask us
to *choose* which one we want to jump to.

scripts/mtags.in:
    Avoid this annoyance by not generating a tag file entry for function
    symbols whose name is the same as the name of their containing type.