Avoid failure of tests/invalid/foreign_decl_line_number with clang.
The above test case is intended to check that the correct line numbers are
reported for the contents of foreign language pragmas.  The way this is done
(i.e. assuming that context information will be printed out along with the
contents of an #error directive) is GCC specific and clang doesn't implement
that behaviour.

tests/invalid/foreign_decl_line_number.err_exp5:
    Add the expected error output for clang.