Fix a bug that broke mmc --make.
The recent shift of the mode of operation into the globals structure also
shifted the point at which the mode of operation is computed to *before* we
have dealt with any implications among the option set.  In particular,
computing the mode of operation did not take the effect of the
'--invoked-by-mmc-make' on the '--make' option into account.

compiler/op_mode.m:
     Account for the value of '--invoked-by-mmc-make' when computing the
     mode of operation.