Speed up the parser a bit.
One speedup is from inlining parse_simple_term_2 into parse_simple_term,
which allows us to avoid a redundant test, and the overhead of passing
around a success flag.

Another speedup is from the fact that we can now discover that a term
is not followed by curried arguments *without* the need for a call.

Give some predicates and variables more meaningful names.