Fix error messages related to bad .module_dep files.
This change makes it a little easier to understand what's going on when the
compiler doesn't understand your .module_dep files, for example when you
bootstrapped with a compiler that's too old.

The error message I got last week was:

    $ mmc --make hello --rebuild
    ** Error: error reading file `pretty_printer.m' to generate dependencies.
    ** Module `pretty_printer' is imported or included by module `list'.
    mercury_compile: cannot find source for module `pretty_printer' in
    directories .

Now the error message is:

    $ mmc --make hello --rebuild
    Making Mercury/int3s/hello.int3
    ** Error reading file
    `/tmp/test2/lib/mercury/ints/Mercury/module_deps/pretty_printer.module_dep':
    failed to parse term ...rebuilding
    ** Error reading file `pretty_printer.m' to generate dependencies.
    ** Module `pretty_printer' is imported or included by module `char'.
    mercury_compile: cannot find source for module `pretty_printer' in
    directories
      '.'

You normally only get this information (about the .module_dep) file if you use
--debug-make.  So this is much more helpful.

When using --debug-make there is now a more helpful message when a
.module_dep file is not found, before the compiler potentially encounters a
later error with a .m file that the user didn't expect to need to provide.

compiler/make.module_dep_file.m:
    As above.

compiler/prog_io_find.m:
    Improve the final error message in the examples above by adding quotes
    around directory names.