Fix a performance problem in mode analysis.
When working on the diff to make_hlds_passes.m, the compiler took effectively
forever on an intermediate version; I killed it after about half an hour
of CPU time. Invoking a debug version of the compiler on that version
of make_hlds_passes.m and interrupting it several times while it was
twiddling its thumbs showed that it was spending its time in computing
a set of candidate vars for auto-initialization. Since the predicate
it was analyzing at the time, do_parse_tree_to_hlds, is quite big and has
lots of variables, the sets of candidate variables could be quite big,
and may have to be computed lots of times. However, all those computations
are wasted, since once the set of candidate vars is computed, mode analysis
checks if the variables in those sets have a solver type that has auto
initialization. In most predicates including this one, NO variable has
such a type, so this test will fail.

compiler/modecheck_conj.m:
    Test whether ANY variable in the procedure being analyzed has a solver
    type with auto initialization. If not, don't bother to compute the
    candidate vars.

compiler/mode_info.m:
    When initializing the mode_info before starting to analyze a procedure,
    record whether any its variables has a solver type with auto
    initialization.

There is no test case for this performance bug. First, I don't understand
the algorithm that computes the candidate vars set well enough to hand craft
a small test that nevertheless tickles this worst-case behavior. Second,
trying to minimize the actual problem code would probably yield something
that is still way bigger than one of our usual test cases. And third,
if the test case fails, it would do so by using huge amounts of CPU time
*without* running out of stack, which is a pretty bad failure mode,
and I have no idea how to impose a CPU time limit on a test case
that would also work on Windows. (The code in question is nondet, and
every backtrack to a choice point frees the det stack used up till then.)

However, a hand test on the actual problem code (which is a snapshot
of the offending version of the entire compiler directory) does show
that this diff fixes the problem.