Improve warnings generated by --warn-non-tail-recursive
Do not warn about non tail recursive calls that are obviously not tail
recursive.  For example:

    qsort([], []).
    qsort([Pivot | T], Left ++ [Pivot | Right]) :-
        partition(Pivot, T, [], Left0, [], Right0),
        qsort(Left0, Left),
        qsort(Right0, Right).

Warn only about the second call to qsort, the first is obviously not tail
recursive.  The second call is not tail recursive because of the call to ++ in
the clause head.

compiler/ml_tailcall.m:
    As we traverse the MLDS backwards, if we encounter a recursive call
    (whether it's a tailcall or not) don't create a warning for subsequent
    tailcalls.