Don't generate near-duplicate tags.
Lots of Mercury types have a name that is also the name of one of the type's
function symbols, like this:

:- type allow_lco
    --->    do_not_allow_lco
    ;       allow_lco.

When jumping to a tag that has two definitions, the type name and one of that
type's function symbols, it doesn't really matter which one we jump to, since
they are next to each other. It is therefore annoying having vim ask us
to *choose* which one we want to jump to.

A while ago, I added code to mtags to fix this annoyance, but it worked
only when the functor had no arguments.

scripts/mtags.in:
    Make the fix work even when the functor has arguments, by making the
    test work on a version of the tag AFTER, not BEFORE, the argument list
    following non-zero-arity function symbols is stripped off.