Add the compiler option --warn-interface-imports-in-parents.
When enabled, this asks the compiler to apply the same rules to parent
modules as it does to every other module, and generate warnings for any
import in the interface of a module if that import is not used in the interface
of that module. By default, we still supress all such warnings on the basis
that a module A imported in the interface of module B may be used in the
interface of module B.C even if it isn't used in B itself, but this option
provides a way to override that presumption.

compiler/options.m:
doc/user_guide.texi:
    Add the option.

compiler/module_qual.qualify_items.m:
    If this option is set, then don't set to empty the set of not-yet-seen-
    to-be-used imports when we see an include_module declaration.

compiler/module_qual.m:
    Add a field to the mq_info (actually, its rarely used part, mq_sub_info)
    that says whether the option is set, for module_qual.qualify_items.m
    to use.

compiler/module_qual.m:
compiler/module_qual.qual_errors.m:
    Have module_qual.m implement the --warn-interface-imports option
    just once, to save having module_qual.qual_errors.m having to implement
    it once per unused module.