Fix bug #388.
Doing 'sudo make install' on Linux with Java 7 or later results in the installed
standard library jars being owned by the root user (which is fine) and with
permissions 0600 (which is not).  This is occurring because as part of building
the library jars we create a jar index using the 'jar -i' command.  Doing so
creates a temporary file with the above permissions and the installed version
inherits those.  Building the index is only useful for executable jars anyway
and the Mercury compiler never builds those, so the fix is to simply not build
the index.

This fix is due to Sebastian Godelet.

compiler/compile_target_code.m:
     As above.