Fix a bug in dead procedure elimination.
If predicate p is not called from Mercury code by it is exported
to foreign language code, we used to suppress warnings about it
being unused, but we did not consider it reachable. This meant that
if another predicate g was called ONLY from p, then we would (a)
generate a warning about g being unused, and (b) delete g from the HLDS.
The latter was a bug.

compiler/dead_proc_elim.m:
    Treat procedures that are exported to foreign language code as being
    just as reachable as procedures that are exported from the module.

    Document two situations in which we do still generate undeserved warnings
    about procedures being unused.