Support MS VC 2015 32-bit
Support compiling Mercury with MS Visual C compiler 2015.
Due to a change on how the compiler supports the C99 runtime functions,
the autoconfigure macros do not work with this version of the compiler.
This change uses the AC_CHECK_DECL macro instead to check for the
C99 (v)snprintf function family.

Configure.ac:
    Introduce mercury_check_for_stdio_functions and
    check for declaration of the (v)snprintf functions.