Improve error messages for invalid determinism categories.
For declarations with an invalid determinism category, for example

    :- pred report_usage_error(io::di, io::uo) is et.

the error message generated by the compiler currently contains all of the term
after the `:- pred' bit.

     Error: invalid determinism category
        `(report_usage_error((io :: di), (io :: uo)) is et)'.

Change this error message so that (1) we report the kind of item that the error
is occurring in and (2) we only report the erroneous part of the term (`et' in
the above example).  Furthermore, make the error message distinguish between
predicate, function and mode declarations that occur at the top-level and those
that occur for type class methods.  Also apply this distinction to the error
messages generated when the inst term of a `with_inst` annotation is invalid.

compiler/parse_item.m:
     Make the above change to error messages generated for invalid determinism
     categories.

     Thread a value through the code that parses declarations that indicates
     whether the declaration is nested inside a type class declaration or not.

     Use that information to generate more specific error messages.

tests/invalid/Mmakefile:
tests/invalid/bad_detism_category.{m,err_exp}:
tests/invalid/bad_with_inst.{m,err_exp}:
     Add test cases for the new error messages.  (Note that none of the
     existing invalid tests covered these ones.)