Fix a problem with linking to libhwloc in high-level C grades
When libhwloc is available it is used by the low level C grades to determine
the number of processors available.  I recently added similar support to the
high level C grades, but I neglected to add -lhwloc to the compiler's command
line in these grades.  This patch fixes that.

compiler/compile_target_code.m:
    Link to libhwloc in all parallel grades.

configure.ac:
    Use the correct name for all instances of the same variable.