Improve library/pretty_printer.m.
Put related types and predicates together. Put the groups of related types
and predicates into a meaningful order. Put the definitions in the
implementation section into the same order as their declarations in the
interface section.

Improve the comments on some of the exported types, functions and predicates.

Add some notes about how the interface could be improved, without actually
changing the interface (for now).

Change the representation of the indent stack to a bespoke type. (Doing this
takes *fewer* lines than properly documenting the old setup.)

Convert some code that used ^elem notation to use plain predicate calls
instead.

Rename an internal predicate to avoid unnecessary ambiguity.

Turn an internal function into a predicate to simplify calls to it.