Don't generate the same error message twice.
When a file is expected to contain module X but actually contains module Y,
we used to generate a message about this error twice: once when processing
the ':- module' marker at the start of the source file, and once just after
the whole source file has been parsed. The two error messages had the same
text, but one was conditional on the warn_wrong_module_name, and the other
was unconditional. Error_util.m cleaned up the duplication, but not generating
duplication is a better fix.

compiler/parse_module.m:
    Make the error_spec for this error always conditional on the
    warn_wrong_module_name option. Document why we do this.

compiler/read_modules.m:
    Don't check for the wrong module name after the whole file has been parsed;
    any such error would have been caught when the ':- module' marker was
    processed.

compiler/error_util.m:
    Note that we no longer need a workaround for the error message duplication
    fixed by this diff.