Fix sparse_bitset.divide_by_set.
The helper predicate `divide_bits_by_set' compares two bit vectors
(words) at the same offset, but it was incorrectly passed an absolute
offset instead of the offset *within* the two words it compares.
That led a call to unchecked_left_shift:

        OffsetBit = unchecked_left_shift(1, Offset)

which compiles to C:

        OffsetBit = 1 << Offset;

whose behaviour is undefined when Offset >= bits_per_int.

The correct expression should be

        OffsetBit = 1 << OffsetInWord;

(strictly speaking, this still relies on undefined behaviour because we
may shift into the sign bit)

It so happens that the incorrect code produces expected value of OffsetBit
on x86 and x86_64 because the left shift instructions mask the shift count
to the lower 5 or 6 bits (for 32-bit and 64-bit words, respectively),
and because the base offset of each bitset_elem is always a multiple
of bits_per_int.

The incorrect code also happens to work when targeting Java and C#
because those languages define their left shift operators to mask the
shift count in the same way.

library/sparse_bitset.m:
    As above.